Move GraphItem to drawables package as Strain
